# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * YouTube Downloader API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface VideoFormat {
  itag: string;
  quality: string;
  qualityLabel: string;
  mimeType: string;
  hasAudio: boolean;
  hasVideo: boolean;
  container: string;
  /** @nullable */
  filesize?: number | null;
}

export interface VideoInfo {
  videoId: string;
  title: string;
  author: string;
  thumbnailUrl: string;
  durationSeconds: number;
  /** @nullable */
  viewCount?: number | null;
  formats: VideoFormat[];
}

export interface PlaylistVideo {
  videoId: string;
  title: string;
  author: string;
  thumbnailUrl: string;
  /** @nullable */
  durationSeconds?: number | null;
  url: string;
}

export interface PlaylistInfo {
  playlistId: string;
  title: string;
  author: string;
  videoCount: number;
  videos: PlaylistVideo[];
}

export interface ErrorResponse {
  error: string;
}

export type GetVideoInfoParams = {
url: string;
};

export type GetPlaylistInfoParams = {
url: string;
};