# lib/api-zod/src/generated/types/playlistInfo.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * YouTube Downloader API
 * OpenAPI spec version: 0.1.0
 */
import type { PlaylistVideo } from './playlistVideo';

export interface PlaylistInfo {
  playlistId: string;
  title: string;
  author: string;
  videoCount: number;
  videos: PlaylistVideo[];
}