# lib/api-zod/src/generated/types/playlistVideo.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * YouTube Downloader API
 * OpenAPI spec version: 0.1.0
 */

export interface PlaylistVideo {
  videoId: string;
  title: string;
  author: string;
  thumbnailUrl: string;
  /** @nullable */
  durationSeconds?: number | null;
  url: string;
}