# lib/api-zod/src/generated/types/videoFormat.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * YouTube Downloader API
 * OpenAPI spec version: 0.1.0
 */

export interface VideoFormat {
  itag: string;
  quality: string;
  qualityLabel: string;
  mimeType: string;
  hasAudio: boolean;
  hasVideo: boolean;
  container: string;
  /** @nullable */
  filesize?: number | null;
}