# lib/api-zod/src/generated/types/videoInfo.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * YouTube Downloader API
 * OpenAPI spec version: 0.1.0
 */
import type { VideoFormat } from './videoFormat';

export interface VideoInfo {
  videoId: string;
  title: string;
  author: string;
  thumbnailUrl: string;
  durationSeconds: number;
  /** @nullable */
  viewCount?: number | null;
  formats: VideoFormat[];
}